requirements.txt
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.4.0
blinker==1.9.0
branca==0.8.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
executing==2.2.1
fastapi==0.120.0
folium==0.20.0
gitdb==4.0.12
gitpython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.9.1
markupsafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.9.0
nest-asyncio==1.6.0
numpy==2.3.4
openai==2.6.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
prompt-toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.2
ptyprocess==0.7.0
pulp==3.3.0
pure-eval==0.2.3
pyarrow==22.0.0
pydantic==2.11.10
pydantic-core==2.33.2
pydeck==0.9.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.48.0
streamlit==1.50.0
streamlit-folium==0.25.3
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
xyzservices==2025.4.0
